+++ public/service-worker.js
@@ -1,1 +1,1 @@
-const CACHE = 'valorae-proxy-server-v21-12-336';
+const CACHE = 'valorae-proxy-server-v21-12-337';